+++ multiply_by_5operator.py
